backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-multipart
python-dotenv
cloudinary
google-generativeai
requests
